[asin] improve asin_taylor1 for tiny inputs
When Re(z) and Im(z) were tiny, both the real and imaginary parts of asin(z)
round to Re(z) and Im(z), but mpfr_can_round did fail since it could not
decide the ternary flag (for rounding to nearest). Since we know the sign of
the error term, we can optimize this case.